backend/requirements.txt
# Backend dependencies 
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.23
asyncpg>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.3
fastapi-users[sqlalchemy]>=12.1.2
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
bcrypt>=4.0.1 
boto3>=1.28.0
langchain>=0.0.335
replicate>=0.21.1
python-dotenv>=1.0.0
boto3-stubs[textract]>=1.28.0
aiofiles>=23.2.1
python-magic>=0.4.27 